requirements.txt
streamlit
pandas
sqlalchemy
fpdf
google-generativeai